chore: rework source code organization
  - Move and rename Pools related stuff under a 'columns' module; in
    anticipation of other columns to come.
  - Do the same for UTxO
  - Factor out specific db logic for the management of those columns
    under their respective modules.
  - Split DiffEpochReg and DiffState out of the 'state' module, into
    their own submodules.
  - Bump iter_borrow up, since it's not technically store-specific.
