+++ amaru/src/iter/borrow.rs
